refactor(maintenance): consolidate local tooling and value guards
