+++ src/main/notebook/host-model-service.ts
@@ -17,7 +17,7 @@ import {
   type RestrictedInferenceResult,
   type RestrictedInferenceRunner
 } from '../acp/restricted-inference-runner'
-import { isRecord } from './value-guards'
+import { isRecord } from '../value-guards'
 
 const MAX_PROMPT_BYTES = 64 * 1024
 const MAX_BATCH_ITEMS = 32
