+++ src/main/settings/connector-template.ts
@@ -13,7 +13,7 @@ import { CONNECTOR_RESOURCE_LIMITS } from './connector-resource-limits'
 import { CONNECTOR_TEMPLATE_MAX_BYTES } from '../../shared/settings'
 import { isCustomConnectorName, toCustomConnectorName } from '../../shared/custom-connector'
 import { normalizeLoopbackOAuthRedirectUri } from '../../shared/oauth-redirect'
-import { isRecord } from './value-guards'
+import { isRecord } from '../value-guards'
 
 export type ConnectorTemplateSource = {
   id: string
