--- src/main/settings/value-guards.ts
@@ -1,2 +1,0 @@
-export const isRecord = (value: unknown): value is Record<string, unknown> =>
-  typeof value === 'object' && value !== null && !Array.isArray(value)
